requirements.txt
fastapi==0.115.12
pydantic==2.11.1
pydantic_settings==2.8.1
python-dotenv==1.1.0
